+++ packages/plugin-network-breadcrumbs/test/network-breadcrumbs.test.js
@@ -27,6 +27,7 @@ XMLHttpRequest.prototype.removeEventListener = function (evt, listener) {
   if (listener === this._listeners[evt]) delete this._listeners[evt]
 }
 
+// mock fetch
 function fetch (url, options, fail, status) {
   return new Promise((resolve, reject) => {
     if (fail) {
@@ -35,6 +36,12 @@ function fetch (url, options, fail, status) {
       resolve({ status })
     }
   })
+}
+
+// mock (fetch) Request
+function Request (url, opts) {
+  this.url = url
+  this.method = (opts && opts.method) || 'GET'
 }
 
 describe('plugin: network breadcrumbs', () => {
@@ -173,6 +180,208 @@ describe('plugin: network breadcrumbs', () => {
         metaData: {
           status: 200,
           request: 'GET /'
+        }
+      }))
+      done()
+    })
+  })
+
+  it('should handle a fetch(url, { method: null })', (done) => {
+    const window = { XMLHttpRequest, fetch }
+
+    const client = new Client(VALID_NOTIFIER)
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa' })
+    client.configure()
+    client.use(plugin, () => [], window)
+
+    window.fetch('/', { method: null }, false, 405).then(() => {
+      expect(client.breadcrumbs.length).toBe(1)
+      expect(client.breadcrumbs[0]).toEqual(jasmine.objectContaining({
+        type: 'request',
+        name: 'fetch() failed',
+        metaData: {
+          status: 405,
+          request: 'null /'
+        }
+      }))
+      done()
+    })
+  })
+
+  it('should handle a fetch() request supplied with a Request object', (done) => {
+    const window = { XMLHttpRequest, fetch }
+
+    const client = new Client(VALID_NOTIFIER)
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa' })
+    client.configure()
+    client.use(plugin, () => [], window)
+
+    const request = new Request('/')
+
+    window.fetch(request, {}, false, 200).then(() => {
+      expect(client.breadcrumbs.length).toBe(1)
+      expect(client.breadcrumbs[0]).toEqual(jasmine.objectContaining({
+        type: 'request',
+        name: 'fetch() succeeded',
+        metaData: {
+          status: 200,
+          request: 'GET /'
+        }
+      }))
+      done()
+    })
+  })
+
+  it('should handle a fetch() request supplied with a Request object that has a method specified', (done) => {
+    const window = { XMLHttpRequest, fetch }
+
+    const client = new Client(VALID_NOTIFIER)
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa' })
+    client.configure()
+    client.use(plugin, () => [], window)
+
+    const request = new Request('/', { method: 'PUT' })
+
+    window.fetch(request, {}, false, 200).then(() => {
+      expect(client.breadcrumbs.length).toBe(1)
+      expect(client.breadcrumbs[0]).toEqual(jasmine.objectContaining({
+        type: 'request',
+        name: 'fetch() succeeded',
+        metaData: {
+          status: 200,
+          request: 'PUT /'
+        }
+      }))
+      done()
+    })
+  })
+
+  it('should handle fetch(null)', (done) => {
+    const window = { XMLHttpRequest, fetch }
+
+    const client = new Client(VALID_NOTIFIER)
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa' })
+    client.configure()
+    client.use(plugin, () => [], window)
+
+    window.fetch(null, {}, false, 404).then(() => {
+      expect(client.breadcrumbs.length).toBe(1)
+      expect(client.breadcrumbs[0]).toEqual(jasmine.objectContaining({
+        type: 'request',
+        name: 'fetch() failed',
+        metaData: {
+          status: 404,
+          request: 'GET null'
+        }
+      }))
+      done()
+    })
+  })
+
+  it('should handle fetch(url, null)', (done) => {
+    const window = { XMLHttpRequest, fetch }
+
+    const client = new Client(VALID_NOTIFIER)
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa' })
+    client.configure()
+    client.use(plugin, () => [], window)
+
+    window.fetch('/', null, false, 200).then(() => {
+      expect(client.breadcrumbs.length).toBe(1)
+      expect(client.breadcrumbs[0]).toEqual(jasmine.objectContaining({
+        type: 'request',
+        name: 'fetch() succeeded',
+        metaData: {
+          status: 200,
+          request: 'GET /'
+        }
+      }))
+      done()
+    })
+  })
+
+  it('should handle fetch(undefined)', (done) => {
+    const window = { XMLHttpRequest, fetch }
+
+    const client = new Client(VALID_NOTIFIER)
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa' })
+    client.configure()
+    client.use(plugin, () => [], window)
+
+    window.fetch(undefined, {}, false, 404).then(() => {
+      expect(client.breadcrumbs.length).toBe(1)
+      expect(client.breadcrumbs[0]).toEqual(jasmine.objectContaining({
+        type: 'request',
+        name: 'fetch() failed',
+        metaData: {
+          status: 404,
+          request: 'GET undefined'
+        }
+      }))
+      done()
+    })
+  })
+
+  it('should handle a fetch(request, { method })', (done) => {
+    const window = { XMLHttpRequest, fetch }
+
+    const client = new Client(VALID_NOTIFIER)
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa' })
+    client.configure()
+    client.use(plugin, () => [], window)
+
+    window.fetch(new Request('/foo', { method: 'GET' }), { method: 'PUT' }, false, 200).then(() => {
+      expect(client.breadcrumbs.length).toBe(1)
+      expect(client.breadcrumbs[0]).toEqual(jasmine.objectContaining({
+        type: 'request',
+        name: 'fetch() succeeded',
+        metaData: {
+          status: 200,
+          request: 'PUT /foo'
+        }
+      }))
+      done()
+    })
+  })
+
+  it('should handle a fetch(request, { method: null })', (done) => {
+    const window = { XMLHttpRequest, fetch }
+
+    const client = new Client(VALID_NOTIFIER)
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa' })
+    client.configure()
+    client.use(plugin, () => [], window)
+
+    window.fetch(new Request('/foo'), { method: null }, false, 405).then(() => {
+      expect(client.breadcrumbs.length).toBe(1)
+      expect(client.breadcrumbs[0]).toEqual(jasmine.objectContaining({
+        type: 'request',
+        name: 'fetch() failed',
+        metaData: {
+          status: 405,
+          request: 'null /foo'
+        }
+      }))
+      done()
+    })
+  })
+
+  it('should handle a fetch(request, { method: undefined })', (done) => {
+    const window = { XMLHttpRequest, fetch }
+
+    const client = new Client(VALID_NOTIFIER)
+    client.setOptions({ apiKey: 'aaaa-aaaa-aaaa-aaaa' })
+    client.configure()
+    client.use(plugin, () => [], window)
+
+    window.fetch(new Request('/foo'), { method: undefined }, false, 200).then(() => {
+      expect(client.breadcrumbs.length).toBe(1)
+      expect(client.breadcrumbs[0]).toEqual(jasmine.objectContaining({
+        type: 'request',
+        name: 'fetch() succeeded',
+        metaData: {
+          status: 200,
+          request: 'GET /foo'
         }
       }))
       done()